댓글 달고 삭제 기능 구현

+++ posts/views.py
@@ -1,16 +1,21 @@
 from django.shortcuts import render,redirect
-from .forms import PostForm, ImageForm
-from .models import Post
+from .forms import PostForm, ImageForm, CommentForm
+from .models import Post,Comment
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST
+
 
 # Create your views here.
 def list(request):
     posts = Post.objects.all()
-    return render(request, 'posts/list.html',{'posts':posts})
+    comment_form = CommentForm() #댓글 폼 보여주기
+    return render(request, 'posts/list.html',{'posts':posts,'comment_form':comment_form})
 
 #로그인을 하지 않으면 접근할 수 없다.
-@login_required    
+@login_required   
+@require_POST #포스트 방식이 아니면 튕겨낸다.
 def create(request):
+    #get 작성 폼 보여줄게post 작성한 내용 저장해줄게
     #1. get방식으로 데이터를 입력할 form을 요청한다.
     #4. 사용자가 데이터를 입력해서 post방식으로 요청한다.
     #9.사용자가 다시 적절한 데이터를 입력해서 post방식으로 요청한다.
@@ -71,6 +76,7 @@ def create(request):
 @login_required
 def update(request,id):
     post = Post.objects.get(id=id)
+    #get 작성 폼 보여줄게post 작성한 내용 저장해줄게
     if post.user == request.user: # 지금 로그인한 사람이 게시물을 작성해따
         if request.method == "POST":
             form = PostForm(request.POST,instance=post)
@@ -87,7 +93,44 @@ def update(request,id):
     else:
         return redirect("posts:list")
 # 삭제하기
+@login_required
 def delete(request,id):
+    # 지우고싶은 게시물 찾기
     post = Post.objects.get(id=id)
-    post.delete()
+    # 로그인한 유저랑 게시물 작성유저랑 같은지 확인
+    if post.user == request.user:
+        post.delete()
+        return redirect("posts:list")
+    else:
+        return redirect("posts:list")
+    
+    
+#댓글 달기
+@login_required
+def comment_create(request, post_id):
+    #댓글 작성 폼을 list에서 보여줄거임.
+    if request.method == 'POST':
+        comment_form = CommentForm(request.POST)
+        # 폼에 데이터 잘 들어갔는지 검증, save
+        if comment_form.is_valid():
+            comment = comment_form.save(commit= False) #유저정보,몇번 글 넣어야하니까 기다려
+            #유저정보
+            comment.user = request.user
+            #몇번 글에 작성할건지
+            comment.post = Post.objects.get(id=post_id)#포스트 번호
+            comment.save() # auto_now_add 를 이미 줘서 그냥 저장됨.
+            return redirect('posts:list')
+    # else:
+    #     pass 필요없음
+
+#댓글 삭제하기\
+# 삭제하기
+@login_required
+def comment_delete(request,post_id,comment_id):
+    # 코멘트
+    comment = Comment.objects.get(id=comment_id)
+    #코멘트를 작성한 유저랑 요청 유저랑 같은지 확인.
+    if comment.user == request.user:
+        comment.delete()
     return redirect("posts:list")
+
